# frontend/cinebook/src/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Globsyn Theatres API
  description: |-
    API for managing Globsyn Theatres online ticket booking system

    Contacts:
    - Adrita Chatterjee: adritachatterji@gmail.com
    - Anamitra Chandra:
    - Souraajit Samanta:
    - Srinjay Das Gupta: dasguptasrinjay2004@gmail.com, https://www.srinjaydg.in
  contact:
    name: Srinjay Das Gupta
    email: dasguptasrinjay2004@gmail.com
    url: https://www.srinjaydg.in
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Development server
paths:
  /movie:
    get:
      summary: Retrieve a list of movies
      tags:
        - Movies
      description: Retrieve a list of movies from the database
      responses:
        '200':
          description: A list of movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
    post:
      summary: Add a new movie to the database
      tags:
        - Movies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movie'
      responses:
        '201':
          description: The movie was successfully added to the database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '500':
          description: Some server error
  /movie/{id}:
    get:
      summary: Retrieve a movie by ID
      tags:
        - Movies
      parameters:
        - $ref: '#/components/parameters/movieId'
      description: Retrieve a list of movies from the database
      responses:
        '200':
          description: The movie with the specified ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '404':
          description: The movie with the specified ID was not found
    put:
      summary: Update a movie by ID
      tags:
        - Movies
      parameters:
        - $ref: '#/components/parameters/movieId'
      description: Update a movie by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movie'
      responses:
        '200':
          description: The movie was successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '404':
          description: The movie with the specified ID was not found
        '500':
          description: Some server error
    delete:
      summary: Delete a movie by ID
      tags:
        - Movies
      parameters:
        - $ref: '#/components/parameters/movieId'
      description: Delete a movie by ID
      responses:
        '200':
          description: The movie was successfully deleted
        '404':
          description: The movie with the specified ID was not found
        '500':
          description: Some server error
...
  /screen:
    get:
      summary: Retrieve a list of screens from the database
      tags:
        - Screens
      description: Get a list of all screens
      responses:
        '200':
          description: A list of screens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Screen'
    post:
      summary: Add a new screen to the database
      tags:
        - Screens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Screen'
      responses:
        '201':
          description: The screen was successfully added to the database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Screen'
        '500':
          description: Some server error
  /screen/{id}:
    get:
      summary: Retrieve a screen by ID
      tags:
        - Screens
      parameters:
        - $ref: '#/components/parameters/screenId'
      description: Retrieve a specific screen by ID
      responses:
        '200':
          description: The screen with the specified ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Screen'
        '404':
          description: The screen with the specified ID was not found
    put:
      summary: Update a screen by ID
      tags:
        - Screens
      parameters:
        - $ref: '#/components/parameters/screenId'
      description: Update a screen by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Screen'
      responses:
        '200':
          description: The screen was successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Screen'
        '404':
          description: The screen with the specified ID was not found
        '500':
          description: Some server error
    delete:
      summary: Delete a screen by ID
      tags:
        - Screens
      parameters:
        - $ref: '#/components/parameters/screenId'
      description: Delete a screen by ID
      responses:
        '200':
          description: The screen was successfully deleted
        '404':
          description: The screen with the specified ID was not found
        '500':
          description: Some server error
...
components:
  schemas:
    Movie:
      type: object
      required:
        - name
        - duration
        - genre
        - rel_date
        - cast
        - director
        - imdb_rating
        - poster
        - language
      properties:
        id:
          type: string
          description: Auto-generated unique identifier
        name:
          type: string
          description: Name of the movie
        duration:
          type: number
          description: Duration of the movie in minutes
        genre:
          type: string
          description: Genre of the movie
        rel_date:
          type: string
          format: date
          description: Release date of the movie
        cast:
          type: array
          items:
            type: string
          description: Cast of the movie
        director:
          type: string
          description: Director of the movie
        imdb_rating:
          type: number
          description: IMDb rating of the movie
        poster:
          type: string
          description: URL of the movie poster
        language:
          type: string
          description: Language of the movie
      example:
        name: The Shawshank Redemption
        duration: 142
        genre: Drama
        rel_date: '1994-09-23'
        cast:
          - Tim Robbins
          - Morgan Freeman
        director: Frank Darabont
        imdb_rating: 9.3
        poster: https://example.com/shawshank_redemption.jpg
        language: English
...
    SeatRange:
      type: object
      required:
        - type
        - startRow
        - startNumber
        - endRow
        - endNumber
      properties:
        type:
          type: string
          description: Type of seat range
        startRow:
          type: string
          description: Starting row of the seat range
        startNumber:
          type: number
          description: Starting seat number of the seat range
        endRow:
          type: string
          description: Ending row of the seat range
        endNumber:
          type: number
          description: Ending seat number of the seat range
      example:
        type: VIP
        startRow: A
        startNumber: 1
        endRow: A
        endNumber: 10
    Screen:
      type: object
      required:
        - screenNumber
        - totalSeats
        - seatLayout
      properties:
        id:
          type: string
          description: Auto-generated unique identifier
        screenNumber:
          type: number
          description: Number of the screen
        totalSeats:
          type: number
          description: Total number of seats
        seatLayout:
          type: array
          items:
            $ref: '#/components/schemas/SeatRange'
        description:
          type: string
          description: Description of the screen
      example:
        screenNumber: 1
        totalSeats: 200
        seatLayout:
          - type: Regular
            startRow: A
            startNumber: 1
            endRow: E
            endNumber: 20
        description: Main screen with regular and VIP seats
...
  parameters:
    movieId:
      name: id
      in: path
      description: ID of the movie
      required: true
      schema:
        type: string
...
    screenId:
      name: id
      in: path
      description: ID of the screen
      required: true
      schema:
        type: string
...